+++ spec/factories/spells.rb
@@ -12,6 +12,10 @@ FactoryBot.define do
       }
     }
 
+    trait :dnd2024 do
+      type { 'Dnd2024::Spell' }
+    end
+
     trait :daggerheart do
       type { 'Daggerheart::Spell' }
       data {